Move TimeSeriesSourceOperator out its factory class
