+++ x-pack/plugin/esql/compute/src/main/java/org/elasticsearch/compute/lucene/TimeSeriesSourceOperator.java
@@ -13,8 +13,6 @@ import org.apache.lucene.index.LeafReaderContext;
 import org.apache.lucene.index.NumericDocValues;
 import org.apache.lucene.index.SortedDocValues;
 import org.apache.lucene.search.DocIdSetIterator;
-import org.apache.lucene.search.Query;
-import org.apache.lucene.search.ScoreMode;
 import org.apache.lucene.search.Scorer;
 import org.apache.lucene.search.Weight;
 import org.apache.lucene.util.BytesRef;
@@ -27,7 +25,6 @@ import org.elasticsearch.compute.data.IntVector;
 import org.elasticsearch.compute.data.LongVector;
 import org.elasticsearch.compute.data.OrdinalBytesRefVector;
 import org.elasticsearch.compute.data.Page;
-import org.elasticsearch.compute.operator.DriverContext;
 import org.elasticsearch.compute.operator.SourceOperator;
 import org.elasticsearch.core.Releasable;
 import org.elasticsearch.core.Releasables;
@@ -41,352 +38,277 @@ import java.io.IOException;
 import java.io.UncheckedIOException;
 import java.util.Arrays;
 import java.util.List;
-import java.util.function.Function;
 
-/**
- * Creates a source operator that takes advantage of the natural sorting of segments in a tsdb index.
- * <p>
- * This source operator loads the _tsid and @timestamp fields, which is used for emitting documents in the correct order. These field values
- * are included in the page as seperate blocks and downstream operators can make use of these loaded time series ids and timestamps.
- * <p>
- * The source operator includes all documents of a time serie with the same page. So the same time series never exists in multiple pages.
- * Downstream operators can make use of this implementation detail.
- * <p>
- * This operator currently only supports shard level concurrency. A new concurrency mechanism should be introduced at the time serie level
- * in order to read tsdb indices in parallel.
- */
-public class TimeSeriesSortedSourceOperatorFactory extends LuceneOperator.Factory {
+public final class TimeSeriesSourceOperator extends SourceOperator {
 
-    private final int maxPageSize;
     private final boolean emitDocIds;
-    private final List<ValuesSourceReaderOperator.FieldInfo> fieldsToExact;
+    private final int maxPageSize;
+    private final BlockFactory blockFactory;
+    private final LuceneSliceQueue sliceQueue;
+    private int currentPagePos = 0;
+    private int remainingDocs;
+    private boolean doneCollecting;
 
-    private TimeSeriesSortedSourceOperatorFactory(
-        List<? extends ShardContext> contexts,
+    private LongVector.Builder timestampsBuilder;
+    private TsidBuilder tsHashesBuilder;
+    private SegmentsIterator iterator;
+    private final List<ValuesSourceReaderOperator.FieldInfo> fieldsToExtracts;
+    private ShardLevelFieldsReader fieldsReader;
+    private DocIdCollector docCollector;
+
+    TimeSeriesSourceOperator(
+        BlockFactory blockFactory,
         boolean emitDocIds,
-        List<ValuesSourceReaderOperator.FieldInfo> fieldsToExact,
-        Function<ShardContext, Query> queryFunction,
-        int taskConcurrency,
+        List<ValuesSourceReaderOperator.FieldInfo> fieldsToExtract,
+        LuceneSliceQueue sliceQueue,
         int maxPageSize,
         int limit
     ) {
-        super(
-            contexts,
-            queryFunction,
-            DataPartitioning.SHARD,
-            query -> { throw new UnsupportedOperationException("locked to SHARD partitioning"); },
-            taskConcurrency,
-            limit,
-            false,
-            ScoreMode.COMPLETE_NO_SCORES
-        );
         this.maxPageSize = maxPageSize;
+        this.blockFactory = blockFactory;
+        this.fieldsToExtracts = fieldsToExtract;
         this.emitDocIds = emitDocIds;
-        this.fieldsToExact = fieldsToExact;
+        this.remainingDocs = limit;
+        this.timestampsBuilder = blockFactory.newLongVectorBuilder(Math.min(limit, maxPageSize));
+        this.tsHashesBuilder = new TsidBuilder(blockFactory, Math.min(limit, maxPageSize));
+        this.sliceQueue = sliceQueue;
     }
 
     @Override
-    public SourceOperator get(DriverContext driverContext) {
-        return new Impl(driverContext.blockFactory(), emitDocIds, fieldsToExact, sliceQueue, maxPageSize, limit);
+    public void finish() {
+        this.doneCollecting = true;
     }
 
     @Override
-    public String describe() {
-        return "TimeSeriesSortedSourceOperator[maxPageSize = " + maxPageSize + ", limit = " + limit + "]";
+    public boolean isFinished() {
+        return doneCollecting;
     }
 
-    public static TimeSeriesSortedSourceOperatorFactory create(
-        int limit,
-        int maxPageSize,
-        int taskConcurrency,
-        boolean emitDocIds,
-        List<? extends ShardContext> contexts,
-        List<ValuesSourceReaderOperator.FieldInfo> fieldsToExact,
-        Function<ShardContext, Query> queryFunction
-    ) {
-        return new TimeSeriesSortedSourceOperatorFactory(
-            contexts,
-            emitDocIds,
-            fieldsToExact,
-            queryFunction,
-            taskConcurrency,
-            maxPageSize,
-            limit
-        );
-    }
-
-    static final class Impl extends SourceOperator {
-
-        private final boolean emitDocIds;
-        private final int maxPageSize;
-        private final BlockFactory blockFactory;
-        private final LuceneSliceQueue sliceQueue;
-        private int currentPagePos = 0;
-        private int remainingDocs;
-        private boolean doneCollecting;
-
-        private LongVector.Builder timestampsBuilder;
-        private TsidBuilder tsHashesBuilder;
-        private SegmentsIterator iterator;
-        private final List<ValuesSourceReaderOperator.FieldInfo> fieldsToExtracts;
-        private ShardLevelFieldsReader fieldsReader;
-        private DocIdCollector docCollector;
-
-        Impl(
-            BlockFactory blockFactory,
-            boolean emitDocIds,
-            List<ValuesSourceReaderOperator.FieldInfo> fieldsToExtract,
-            LuceneSliceQueue sliceQueue,
-            int maxPageSize,
-            int limit
-        ) {
-            this.maxPageSize = maxPageSize;
-            this.blockFactory = blockFactory;
-            this.fieldsToExtracts = fieldsToExtract;
-            this.emitDocIds = emitDocIds;
-            this.remainingDocs = limit;
-            this.timestampsBuilder = blockFactory.newLongVectorBuilder(Math.min(limit, maxPageSize));
-            this.tsHashesBuilder = new TsidBuilder(blockFactory, Math.min(limit, maxPageSize));
-            this.sliceQueue = sliceQueue;
+    @Override
+    public Page getOutput() {
+        if (isFinished()) {
+            return null;
         }
 
-        @Override
-        public void finish() {
-            this.doneCollecting = true;
+        if (remainingDocs <= 0) {
+            doneCollecting = true;
+            return null;
         }
 
-        @Override
-        public boolean isFinished() {
-            return doneCollecting;
-        }
-
-        @Override
-        public Page getOutput() {
-            if (isFinished()) {
-                return null;
-            }
-
-            if (remainingDocs <= 0) {
-                doneCollecting = true;
-                return null;
-            }
-
-            Page page = null;
-            Block[] blocks = new Block[(emitDocIds ? 3 : 2) + fieldsToExtracts.size()];
-            try {
-                if (iterator == null) {
-                    var slice = sliceQueue.nextSlice();
-                    if (slice == null) {
-                        doneCollecting = true;
-                        return null;
-                    }
-                    Releasables.close(fieldsReader);
-                    fieldsReader = new ShardLevelFieldsReader(blockFactory, slice.shardContext(), fieldsToExtracts);
-                    iterator = new SegmentsIterator(slice);
-                    if (emitDocIds) {
-                        docCollector = new DocIdCollector(blockFactory, slice.shardContext());
-                    }
+        Page page = null;
+        Block[] blocks = new Block[(emitDocIds ? 3 : 2) + fieldsToExtracts.size()];
+        try {
+            if (iterator == null) {
+                var slice = sliceQueue.nextSlice();
+                if (slice == null) {
+                    doneCollecting = true;
+                    return null;
                 }
+                Releasables.close(fieldsReader);
+                fieldsReader = new ShardLevelFieldsReader(blockFactory, slice.shardContext(), fieldsToExtracts);
+                iterator = new SegmentsIterator(slice);
+                if (emitDocIds) {
+                    docCollector = new DocIdCollector(blockFactory, slice.shardContext());
+                }
+            }
+            if (docCollector != null) {
+                docCollector.prepareForCollecting(Math.min(remainingDocs, maxPageSize));
+            }
+            fieldsReader.prepareForReading(Math.min(remainingDocs, maxPageSize));
+            iterator.readDocsForNextPage();
+            if (currentPagePos > 0) {
+                int blockIndex = 0;
                 if (docCollector != null) {
-                    docCollector.prepareForCollecting(Math.min(remainingDocs, maxPageSize));
+                    blocks[blockIndex++] = docCollector.build().asBlock();
                 }
-                fieldsReader.prepareForReading(Math.min(remainingDocs, maxPageSize));
-                iterator.readDocsForNextPage();
-                if (currentPagePos > 0) {
-                    int blockIndex = 0;
-                    if (docCollector != null) {
-                        blocks[blockIndex++] = docCollector.build().asBlock();
-                    }
-                    blocks[blockIndex++] = tsHashesBuilder.build().asBlock();
-                    tsHashesBuilder = new TsidBuilder(blockFactory, Math.min(remainingDocs, maxPageSize));
-                    blocks[blockIndex++] = timestampsBuilder.build().asBlock();
-                    timestampsBuilder = blockFactory.newLongVectorBuilder(Math.min(remainingDocs, maxPageSize));
-                    System.arraycopy(fieldsReader.buildBlocks(), 0, blocks, blockIndex, fieldsToExtracts.size());
-                    page = new Page(currentPagePos, blocks);
-                    currentPagePos = 0;
+                blocks[blockIndex++] = tsHashesBuilder.build().asBlock();
+                tsHashesBuilder = new TsidBuilder(blockFactory, Math.min(remainingDocs, maxPageSize));
+                blocks[blockIndex++] = timestampsBuilder.build().asBlock();
+                timestampsBuilder = blockFactory.newLongVectorBuilder(Math.min(remainingDocs, maxPageSize));
+                System.arraycopy(fieldsReader.buildBlocks(), 0, blocks, blockIndex, fieldsToExtracts.size());
+                page = new Page(currentPagePos, blocks);
+                currentPagePos = 0;
+            }
+            if (iterator.completed()) {
+                Releasables.close(docCollector, fieldsReader);
+                iterator = null;
+            }
+        } catch (IOException e) {
+            throw new UncheckedIOException(e);
+        } finally {
+            if (page == null) {
+                Releasables.closeExpectNoException(blocks);
+            }
+        }
+        return page;
+    }
+
+    @Override
+    public void close() {
+        Releasables.closeExpectNoException(timestampsBuilder, tsHashesBuilder, docCollector, fieldsReader);
+    }
+
+    class SegmentsIterator {
+        private final PriorityQueue<LeafIterator> mainQueue;
+        private final PriorityQueue<LeafIterator> oneTsidQueue;
+        final LuceneSlice luceneSlice;
+
+        SegmentsIterator(LuceneSlice luceneSlice) throws IOException {
+            this.luceneSlice = luceneSlice;
+            this.mainQueue = new PriorityQueue<>(luceneSlice.numLeaves()) {
+                @Override
+                protected boolean lessThan(LeafIterator a, LeafIterator b) {
+                    return a.timeSeriesHash.compareTo(b.timeSeriesHash) < 0;
                 }
-                if (iterator.completed()) {
-                    Releasables.close(docCollector, fieldsReader);
-                    iterator = null;
-                }
-            } catch (IOException e) {
-                throw new UncheckedIOException(e);
-            } finally {
-                if (page == null) {
-                    Releasables.closeExpectNoException(blocks);
+            };
+            Weight weight = luceneSlice.weight();
+            int maxSegmentOrd = 0;
+            for (var leafReaderContext : luceneSlice.leaves()) {
+                LeafIterator leafIterator = new LeafIterator(weight, leafReaderContext.leafReaderContext());
+                leafIterator.nextDoc();
+                if (leafIterator.docID != DocIdSetIterator.NO_MORE_DOCS) {
+                    mainQueue.add(leafIterator);
+                    maxSegmentOrd = Math.max(maxSegmentOrd, leafIterator.segmentOrd);
                 }
             }
-            return page;
+            this.oneTsidQueue = new PriorityQueue<>(mainQueue.size()) {
+                @Override
+                protected boolean lessThan(LeafIterator a, LeafIterator b) {
+                    return a.timestamp > b.timestamp;
+                }
+            };
         }
 
-        @Override
-        public void close() {
-            Releasables.closeExpectNoException(timestampsBuilder, tsHashesBuilder, docCollector, fieldsReader);
+        // TODO: add optimize for one leaf?
+        void readDocsForNextPage() throws IOException {
+            Thread executingThread = Thread.currentThread();
+            for (LeafIterator leaf : mainQueue) {
+                leaf.reinitializeIfNeeded(executingThread);
+            }
+            for (LeafIterator leaf : oneTsidQueue) {
+                leaf.reinitializeIfNeeded(executingThread);
+            }
+            do {
+                PriorityQueue<LeafIterator> sub = subQueueForNextTsid();
+                if (sub.size() == 0) {
+                    break;
+                }
+                tsHashesBuilder.appendNewTsid(sub.top().timeSeriesHash);
+                if (readValuesForOneTsid(sub)) {
+                    break;
+                }
+            } while (mainQueue.size() > 0);
         }
 
-        class SegmentsIterator {
-            private final PriorityQueue<LeafIterator> mainQueue;
-            private final PriorityQueue<LeafIterator> oneTsidQueue;
-            final LuceneSlice luceneSlice;
-
-            SegmentsIterator(LuceneSlice luceneSlice) throws IOException {
-                this.luceneSlice = luceneSlice;
-                this.mainQueue = new PriorityQueue<>(luceneSlice.numLeaves()) {
-                    @Override
-                    protected boolean lessThan(LeafIterator a, LeafIterator b) {
-                        return a.timeSeriesHash.compareTo(b.timeSeriesHash) < 0;
-                    }
-                };
-                Weight weight = luceneSlice.weight();
-                int maxSegmentOrd = 0;
-                for (var leafReaderContext : luceneSlice.leaves()) {
-                    LeafIterator leafIterator = new LeafIterator(weight, leafReaderContext.leafReaderContext());
-                    leafIterator.nextDoc();
-                    if (leafIterator.docID != DocIdSetIterator.NO_MORE_DOCS) {
-                        mainQueue.add(leafIterator);
-                        maxSegmentOrd = Math.max(maxSegmentOrd, leafIterator.segmentOrd);
-                    }
+        private boolean readValuesForOneTsid(PriorityQueue<LeafIterator> sub) throws IOException {
+            do {
+                LeafIterator top = sub.top();
+                currentPagePos++;
+                remainingDocs--;
+                if (docCollector != null) {
+                    docCollector.collect(top.segmentOrd, top.docID);
                 }
-                this.oneTsidQueue = new PriorityQueue<>(mainQueue.size()) {
-                    @Override
-                    protected boolean lessThan(LeafIterator a, LeafIterator b) {
-                        return a.timestamp > b.timestamp;
-                    }
-                };
-            }
-
-            // TODO: add optimize for one leaf?
-            void readDocsForNextPage() throws IOException {
-                Thread executingThread = Thread.currentThread();
-                for (LeafIterator leaf : mainQueue) {
-                    leaf.reinitializeIfNeeded(executingThread);
+                tsHashesBuilder.appendOrdinal();
+                timestampsBuilder.appendLong(top.timestamp);
+                fieldsReader.readValues(top.segmentOrd, top.docID);
+                if (top.nextDoc()) {
+                    sub.updateTop();
+                } else if (top.docID == DocIdSetIterator.NO_MORE_DOCS) {
+                    sub.pop();
+                } else {
+                    mainQueue.add(sub.pop());
                 }
-                for (LeafIterator leaf : oneTsidQueue) {
-                    leaf.reinitializeIfNeeded(executingThread);
+                if (remainingDocs <= 0 || currentPagePos >= maxPageSize) {
+                    return true;
                 }
-                do {
-                    PriorityQueue<LeafIterator> sub = subQueueForNextTsid();
-                    if (sub.size() == 0) {
-                        break;
-                    }
-                    tsHashesBuilder.appendNewTsid(sub.top().timeSeriesHash);
-                    if (readValuesForOneTsid(sub)) {
-                        break;
-                    }
-                } while (mainQueue.size() > 0);
-            }
+            } while (sub.size() > 0);
+            return false;
+        }
 
-            private boolean readValuesForOneTsid(PriorityQueue<LeafIterator> sub) throws IOException {
-                do {
-                    LeafIterator top = sub.top();
-                    currentPagePos++;
-                    remainingDocs--;
-                    if (docCollector != null) {
-                        docCollector.collect(top.segmentOrd, top.docID);
-                    }
-                    tsHashesBuilder.appendOrdinal();
-                    timestampsBuilder.appendLong(top.timestamp);
-                    fieldsReader.readValues(top.segmentOrd, top.docID);
-                    if (top.nextDoc()) {
-                        sub.updateTop();
-                    } else if (top.docID == DocIdSetIterator.NO_MORE_DOCS) {
-                        sub.pop();
+        private PriorityQueue<LeafIterator> subQueueForNextTsid() {
+            if (oneTsidQueue.size() == 0 && mainQueue.size() > 0) {
+                LeafIterator last = mainQueue.pop();
+                oneTsidQueue.add(last);
+                while (mainQueue.size() > 0) {
+                    var top = mainQueue.top();
+                    if (top.timeSeriesHash.equals(last.timeSeriesHash)) {
+                        oneTsidQueue.add(mainQueue.pop());
                     } else {
-                        mainQueue.add(sub.pop());
+                        break;
                     }
-                    if (remainingDocs <= 0 || currentPagePos >= maxPageSize) {
-                        return true;
-                    }
-                } while (sub.size() > 0);
+                }
+            }
+            return oneTsidQueue;
+        }
+
+        boolean completed() {
+            return mainQueue.size() == 0 && oneTsidQueue.size() == 0;
+        }
+    }
+
+    static class LeafIterator {
+        private final int segmentOrd;
+        private final Weight weight;
+        private final LeafReaderContext leafContext;
+        private SortedDocValues tsids;
+        private NumericDocValues timestamps;
+        private DocIdSetIterator disi;
+        private Thread createdThread;
+
+        private long timestamp;
+        private int lastTsidOrd = -1;
+        private BytesRef timeSeriesHash;
+        private int docID = -1;
+
+        LeafIterator(Weight weight, LeafReaderContext leafContext) throws IOException {
+            this.segmentOrd = leafContext.ord;
+            this.weight = weight;
+            this.leafContext = leafContext;
+            this.createdThread = Thread.currentThread();
+            tsids = leafContext.reader().getSortedDocValues("_tsid");
+            timestamps = DocValues.unwrapSingleton(leafContext.reader().getSortedNumericDocValues("@timestamp"));
+            final Scorer scorer = weight.scorer(leafContext);
+            disi = scorer != null ? scorer.iterator() : DocIdSetIterator.empty();
+        }
+
+        boolean nextDoc() throws IOException {
+            docID = disi.nextDoc();
+            if (docID == DocIdSetIterator.NO_MORE_DOCS) {
                 return false;
             }
+            boolean advanced = timestamps.advanceExact(docID);
+            assert advanced;
+            timestamp = timestamps.longValue();
+            advanced = tsids.advanceExact(docID);
+            assert advanced;
 
-            private PriorityQueue<LeafIterator> subQueueForNextTsid() {
-                if (oneTsidQueue.size() == 0 && mainQueue.size() > 0) {
-                    LeafIterator last = mainQueue.pop();
-                    oneTsidQueue.add(last);
-                    while (mainQueue.size() > 0) {
-                        var top = mainQueue.top();
-                        if (top.timeSeriesHash.equals(last.timeSeriesHash)) {
-                            oneTsidQueue.add(mainQueue.pop());
-                        } else {
-                            break;
-                        }
-                    }
-                }
-                return oneTsidQueue;
-            }
-
-            boolean completed() {
-                return mainQueue.size() == 0 && oneTsidQueue.size() == 0;
+            int ord = tsids.ordValue();
+            if (ord != lastTsidOrd) {
+                timeSeriesHash = tsids.lookupOrd(ord);
+                lastTsidOrd = ord;
+                return false;
+            } else {
+                return true;
             }
         }
 
-        static class LeafIterator {
-            private final int segmentOrd;
-            private final Weight weight;
-            private final LeafReaderContext leafContext;
-            private SortedDocValues tsids;
-            private NumericDocValues timestamps;
-            private DocIdSetIterator disi;
-            private Thread createdThread;
-
-            private long timestamp;
-            private int lastTsidOrd = -1;
-            private BytesRef timeSeriesHash;
-            private int docID = -1;
-
-            LeafIterator(Weight weight, LeafReaderContext leafContext) throws IOException {
-                this.segmentOrd = leafContext.ord;
-                this.weight = weight;
-                this.leafContext = leafContext;
-                this.createdThread = Thread.currentThread();
+        void reinitializeIfNeeded(Thread executingThread) throws IOException {
+            if (executingThread != createdThread) {
                 tsids = leafContext.reader().getSortedDocValues("_tsid");
                 timestamps = DocValues.unwrapSingleton(leafContext.reader().getSortedNumericDocValues("@timestamp"));
                 final Scorer scorer = weight.scorer(leafContext);
                 disi = scorer != null ? scorer.iterator() : DocIdSetIterator.empty();
-            }
-
-            boolean nextDoc() throws IOException {
-                docID = disi.nextDoc();
-                if (docID == DocIdSetIterator.NO_MORE_DOCS) {
-                    return false;
+                if (docID != -1) {
+                    disi.advance(docID);
                 }
-                boolean advanced = timestamps.advanceExact(docID);
-                assert advanced;
-                timestamp = timestamps.longValue();
-                advanced = tsids.advanceExact(docID);
-                assert advanced;
-
-                int ord = tsids.ordValue();
-                if (ord != lastTsidOrd) {
-                    timeSeriesHash = tsids.lookupOrd(ord);
-                    lastTsidOrd = ord;
-                    return false;
-                } else {
-                    return true;
-                }
-            }
-
-            void reinitializeIfNeeded(Thread executingThread) throws IOException {
-                if (executingThread != createdThread) {
-                    tsids = leafContext.reader().getSortedDocValues("_tsid");
-                    timestamps = DocValues.unwrapSingleton(leafContext.reader().getSortedNumericDocValues("@timestamp"));
-                    final Scorer scorer = weight.scorer(leafContext);
-                    disi = scorer != null ? scorer.iterator() : DocIdSetIterator.empty();
-                    if (docID != -1) {
-                        disi.advance(docID);
-                    }
-                    createdThread = executingThread;
-                }
+                createdThread = executingThread;
             }
         }
+    }
 
-        @Override
-        public String toString() {
-            return this.getClass().getSimpleName() + "[" + "maxPageSize=" + maxPageSize + ", remainingDocs=" + remainingDocs + "]";
-        }
-
+    @Override
+    public String toString() {
+        return this.getClass().getSimpleName() + "[" + "maxPageSize=" + maxPageSize + ", remainingDocs=" + remainingDocs + "]";
     }
 
     static class BlockLoaderFactory extends ValuesSourceReaderOperator.DelegatingBlockLoaderFactory {
