+++ x-pack/plugin/esql/compute/src/test/java/org/elasticsearch/compute/lucene/TimeSeriesSortedSourceOperatorTests.java
@@ -239,7 +239,7 @@ public class TimeSeriesSortedSourceOperatorTests extends AnyOperatorTestCase {
             try (var reader = writer.getReader()) {
                 var ctx = new LuceneSourceOperatorTests.MockShardContext(reader, 0);
                 Query query = randomFrom(LongField.newRangeQuery("@timestamp", 0, t0), new MatchNoDocsQuery());
-                var timeSeriesFactory = TimeSeriesSortedSourceOperatorFactory.create(
+                var timeSeriesFactory = TimeSeriesSourceOperatorFactory.create(
                     Integer.MAX_VALUE,
                     randomIntBetween(1, 1024),
                     1,
@@ -324,7 +324,7 @@ public class TimeSeriesSortedSourceOperatorTests extends AnyOperatorTestCase {
 
     }
 
-    public static TimeSeriesSortedSourceOperatorFactory createTimeSeriesSourceOperator(
+    public static TimeSeriesSourceOperatorFactory createTimeSeriesSourceOperator(
         Directory directory,
         Consumer<IndexReader> readerConsumer,
         boolean emitDocIds,
@@ -370,7 +370,7 @@ public class TimeSeriesSortedSourceOperatorTests extends AnyOperatorTestCase {
             )
             .toList();
 
-        return TimeSeriesSortedSourceOperatorFactory.create(limit, maxPageSize, 1, emitDocIds, List.of(ctx), fieldInfos, queryFunction);
+        return TimeSeriesSourceOperatorFactory.create(limit, maxPageSize, 1, emitDocIds, List.of(ctx), fieldInfos, queryFunction);
     }
 
     public static void writeTS(RandomIndexWriter iw, long timestamp, Object[] dimensions, Object[] metrics) throws IOException {
